Purpose
At a young age The Bakers Daughter had to grow up fast. Having to grow up fast was because of the responsibilities The Bakers Daughter was given.

The Bakers Daughter outlet was playing around in the family bakery business.

The Bakers Daughter enjoyed watching the bakery ingredients come to life in the workers specialized areas of the bakery. Eventually, these memories led The Bakers Daughter to open a bakery business for children.

Currently, The Bakers Daughter is eager to share more with you, to teach you some basic life skills that go hand in hand with baking and other matters that are close to The Bakers Daughter heart.

It is my hope to connect with each one of you in some way.